redoing the usage scripts
