+++ scripts/generate-usage-data/example.env
@@ -1,0 +1,3 @@
+# environment variables 
+
+API_TOKEN="<API_Token>"
