+++ scripts/generate-usage-data/requirements.txt
@@ -1,3 +1,4 @@
+python-dotenv
 argparse
 python-dateutil
 requests
